+++ models/rbfn.py
@@ -45,28 +45,28 @@ class RBFN:
 
     def g(self, x):
         return self._g(self.ker, x, self.params)
-    
+
     def obj(self, x):
         return self._obj(self.ker, x, self.params)
 
-    def step(self, x, loop=3, _mask=None):
+    def step(self, x, loop=3, _mask=None, **kwargs):
         for _ in range(loop):
-            value, grads = self._obj(self.ker, x, self.params, mask=_mask)
+            value, grads = self._obj(self.ker, x, self.params, mask=_mask, **kwargs)
             self.opt_state = self.opt_update(self.t, grads, self.opt_state)
         self.t += 1
         return value
-    
-    def step_online(self, x, loop=3):
+
+    def step_online(self, x, loop=3, **kwargs):
         """
         Assuming the role for splitting up data for online learning for now.
         Pad for t=0...self.window.
         """
-        self.mask = index_update(self.mask, self.t - 1, 1.)
+        self.mask = index_update(self.mask, self.t - 1, 1.0)
         if self.t < self.window:
-            z = index_update(np.zeros((self.window, x.shape[1])), index[:self.t + 1, :], x[:self.t + 1, :])
+            z = index_update(np.zeros((self.window, x.shape[1])), index[: self.t + 1, :], x[: self.t + 1, :])
         else:
-            z = x[self.t: self.t + self.window]
-        return self.step(z, loop, self.mask)
+            z = x[self.t : self.t + self.window]
+        return self.step(z, loop, self.mask, **kwargs)
 
     @staticmethod
     @partial(jit, static_argnums=0)
@@ -75,10 +75,11 @@ class RBFN:
         return ker(x, c, σ) @ W - np.exp(-(τ ** 2)) * x  # (4)
 
     @staticmethod
-    def _mse(ker: Callable, x: DeviceArray, p: dict[str, DeviceArray], mask: Optional[DeviceArray]=None):
+    def _mse(ker: Callable, x: DeviceArray, p: dict[str, DeviceArray], mask: Optional[DeviceArray]):
         """||g(x_{t-1}) + x_{t-1} - x_t||²"""
         if mask is None:
             return np.mean(np.square(RBFN._g(ker, x[:-1], p) + x[:-1] - x[1:]))
         return np.sum(np.square(RBFN._g(ker, x[:-1], p) + x[:-1] - x[1:]) * mask) / np.sum(mask)
+
 
 # %%
